- Now we can get and see images preview correctly in space plus near every preview displays square with average color - Uploading images from gallery. Used "requestLegacyExternalStorage" (for Android X+), probably it will deprecate in future, but yet its okay - Minor ui improvements

+++ data/src/main/java/com/netcracker_study_autumn_2020/data/mapper/ImageEntityDtoMapper.java
@@ -41,7 +41,7 @@ public class ImageEntityDtoMapper extends BaseMapper<ImageEntity, ImageDto> {
             e.printStackTrace();
         }
         imageDto.setUserId(o1.getUserId());
-        imageDto.setAverageColor(o1.getAverageColor());
+        imageDto.setAverageColor((int) (o1.getAverageColor()));
         imageDto.setPath(o1.getPath());
         imageDto.setPreviewPath(o1.getPreviewPath());
