+++ domain/src/main/java/com/netcracker_study_autumn_2020/domain/dto/ImageDto.java
@@ -9,7 +9,7 @@ public class ImageDto {
     private String path;
     private String previewPath;
     private Date createTime;
-    private long averageColor;
+    private int averageColor;
     private long size;
     private Date modifiedTime;
 
@@ -53,11 +53,11 @@ public class ImageDto {
         this.createTime = createTime;
     }
 
-    public long getAverageColor() {
+    public int getAverageColor() {
         return averageColor;
     }
 
-    public void setAverageColor(long averageColor) {
+    public void setAverageColor(int averageColor) {
         this.averageColor = averageColor;
     }
